Creating controller for order

+++ src/main/java/com/satyam/SpringEcom/model/dto/OrderItemRequest.java
@@ -1,5 +1,7 @@
 package com.satyam.SpringEcom.model.dto;
 
-public record OrderItemRequest(int productId,int quantity) {
+public record OrderItemRequest(
+        Long productId,
+        int quantity) {
 
 }
