+++ src/main/java/com/satyam/SpringEcom/model/dto/OrderItemResponse.java
@@ -2,6 +2,6 @@ package com.satyam.SpringEcom.model.dto;
 
 import java.math.BigDecimal;
 
-public record OrderItemResponse(String productName, int quantity, BigDecimal totalPrice) {
+public record OrderItemResponse(String productName, int quantity, int totalPrice) {
 
 }
